refactor(metrics): extract trend module
